backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
face-recognition==1.3.0
Pillow==10.4.0
numpy==1.26.4
gunicorn==22.0.0
